code redundancy fixes along with new common.dll
using double in mathematics now

+++ FFXIVAPP.Plugin.Parse/Models/StatGroups/Monster.Stats.Damage.cs
@@ -53,7 +53,7 @@ namespace FFXIVAPP.Plugin.Parse.Models.StatGroups
             var fields = line.GetType()
                              .GetProperties();
 
-            var currentDamage = line.Crit ? line.Amount > 0 ? ParseHelper.GetOriginalAmount(line.Amount, (double) .5) : 0 : line.Amount;
+            var currentDamage = line.Crit ? line.Amount > 0 ? ParseHelper.GetOriginalAmount(line.Amount, .5) : 0 : line.Amount;
             if (currentDamage > 0)
             {
                 ParseHelper.LastAmountByAction.EnsureMonsterAction(line.Source, line.Action, currentDamage);
