+++ FFXIVAPP.Plugin.Parse/Models/StatGroups/Player.Stats.HealingOverHealing.cs
@@ -47,7 +47,7 @@ namespace FFXIVAPP.Plugin.Parse.Models.StatGroups
                 //LineHistory.Add(new LineHistory(line));
             }
 
-            var currentHealing = line.Crit ? line.Amount > 0 ? ParseHelper.GetOriginalAmount(line.Amount, (double) .5) : 0 : line.Amount;
+            var currentHealing = line.Crit ? line.Amount > 0 ? ParseHelper.GetOriginalAmount(line.Amount, .5) : 0 : line.Amount;
             if (currentHealing > 0)
             {
                 ParseHelper.LastAmountByAction.EnsurePlayerAction(line.Source, line.Action, currentHealing);
